requirements.txt
# Core
fastapi==0.115.12
uvicorn==0.34.2
python-dotenv==1.1.0
pydantic==2.11.4

# LangChain + LangGraph
langchain==0.3.25
langchain-google-genai==2.1.4
langgraph==0.4.1
langchain-community==0.3.23
langsmith==0.3.42

# A2A Protocol
a2a-sdk==0.2.9

# Vector DBs
pinecone-client==5.0.1
pgvector==0.3.6

# Database
sqlalchemy==2.0.40
psycopg2-binary==2.9.10
supabase==2.15.1

# Search
google-search-results==2.4.2
serpapi==0.1.5

# Utilities
httpx==0.28.1
python-multipart==0.0.20

google-generativeai==0.8.6

#streamlit
streamlit==1.45.0